SentimentChatbot/requirements.txt
botbuilder-integration-aiohttp>=4.15.0
azure-ai-textanalytics>=5.3.0